# [Merged by Bors] - sql: store ballots, blocks and layers in sqlite 