--- a/crypto/sha3/keccakf.go
+++ b/crypto/sha3/keccakf.go
@@ -2,7 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//  +build !amd64 appengine gccgo
+//go:build !amd64 || appengine || gccgo
+// +build !amd64 appengine gccgo
 
 package sha3