--- a/crypto/sha3/xor.go
+++ b/crypto/sha3/xor.go
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+//go:build (!amd64 && !386 && !ppc64le) || appengine
 // +build !amd64,!386,!ppc64le appengine
 
 package sha3