--- a/go.mod
+++ b/go.mod
@@ -1,7 +1,7 @@
 module github.com/spacemeshos/go-spacemesh
 
 require (
-	crawshaw.io/sqlite v0.3.2
+	crawshaw.io/sqlite v0.3.3-0.20211227050848-2cdb5c1a86a1
 	github.com/ALTree/bigfloat v0.0.0-20201218142103-4a33235224ec
 	github.com/Microsoft/go-winio v0.4.14 // indirect
 	github.com/bits-and-blooms/bitset v1.2.0