--- a/mesh/mesh.go
+++ b/mesh/mesh.go
@@ -13,6 +13,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/database"
 	"github.com/spacemeshos/go-spacemesh/events"
 	"github.com/spacemeshos/go-spacemesh/log"
+	"github.com/spacemeshos/go-spacemesh/sql"
 )
 
 var (
@@ -86,9 +87,6 @@ func NewMesh(db *DB, atxDb AtxDB, trtl tortoise, txPool txMemPool, state state, 
 	gLyr := types.GetEffectiveGenesis()
 	for i := types.NewLayerID(1); !i.After(gLyr); i = i.Add(1) {
 		if i.Before(gLyr) {
-			if err := msh.SetZeroBallotLayer(i); err != nil {
-				msh.With().Panic("failed to set zero-ballot for genesis layer", i, log.Err(err))
-			}
 			if err := msh.SetZeroBlockLayer(i); err != nil {
 				msh.With().Panic("failed to set zero-block for genesis layer", i, log.Err(err))
 			}
@@ -360,6 +358,7 @@ func (msh *Mesh) persistLayerHashes(ctx context.Context, from, to types.LayerID)
 			logger.With().Error("failed to get valid block IDs", i, log.Err(err))
 			return err
 		}
+
 		hash := types.EmptyLayerHash
 		if len(validBlockIDs) > 0 {
 			hash = types.CalcBlocksHash32(validBlockIDs, nil)
@@ -577,7 +576,7 @@ func (msh *Mesh) ProcessLayerPerHareOutput(ctx context.Context, layerID types.La
 		Status:  events.LayerStatusTypeApproved,
 	})
 
-	logger.Info("saving hare output for layer")
+	logger.With().Info("saving hare output for layer")
 	if err := msh.SaveHareConsensusOutput(ctx, layerID, blockID); err != nil {
 		logger.Error("saving layer hare output failed")
 	}
@@ -658,30 +657,6 @@ func uniqueTxIds(blocks []*types.Block, seenTxIds map[types.TransactionID]struct
 
 var errLayerHasBallot = errors.New("layer has ballot")
 
-// SetZeroBallotLayer tags lyr as a layer without ballots.
-func (msh *Mesh) SetZeroBallotLayer(lyr types.LayerID) error {
-	msh.With().Info("tagging zero ballot layer", lyr)
-	// check database for layer
-	if l, err := msh.GetLayer(lyr); err != nil {
-		// database error
-		if !errors.Is(err, database.ErrNotFound) {
-			msh.With().Error("error trying to fetch layer from database", lyr, log.Err(err))
-			return err
-		}
-	} else if len(l.Ballots()) != 0 {
-		// layer exists
-		msh.With().Error("layer has ballots, cannot tag as zero ballot layer",
-			lyr,
-			l,
-			log.Int("num_ballots", len(l.Ballots())))
-		return errLayerHasBallot
-	}
-
-	msh.setLatestLayer(lyr)
-	// layer doesn't exist, need to insert new layer
-	return msh.AddZeroBallotLayer(lyr)
-}
-
 var errLayerHasBlock = errors.New("layer has block")
 
 // SetZeroBlockLayer tags lyr as a layer without blocks.
@@ -690,7 +665,7 @@ func (msh *Mesh) SetZeroBlockLayer(lyr types.LayerID) error {
 	// check database for layer
 	if l, err := msh.GetLayer(lyr); err != nil {
 		// database error
-		if !errors.Is(err, database.ErrNotFound) {
+		if !errors.Is(err, sql.ErrNotFound) {
 			msh.With().Error("error trying to fetch layer from database", lyr, log.Err(err))
 			return err
 		}