@countvonzero 2022-01-12
wow! super happy to move to relational database for data integrity. a couple high level questions

- do you observe any performance degradation?
in one of my previous project we ported geth to use sqlite because we need multi-threaded read and leveldb doesn't support that. my recollection was that there was a performance penalty due to the sheer amount of data.

- wrt to go version, why not go straight to 17? the tidy bug seems to be fixed at that version.