@dshulyak 2022-01-12
> do you observe any performance degradation?

i observed an improvement with more data, and small degradation when there is very little data.
for example for loading 10 000 layers

```
benchmark                             old ns/op       new ns/op       delta
BenchmarkRerun/Verifying/10000-16     15456916985     13337943478     -13.71%
```

but relative improvement will be much larger than -13.71% when encoding will be replaced, as xdr wastes a lot of time.
with 100, 1000 layers leveldb is slightly faster, but i think this is less relevant.

i actually expected sqlite to be faster, as leveldb data structure is write-optimized. @countvonzero do you remember if `database/sql` was used in the previous project or a thin wrapper? there are some explanations why `database/sql` is not a good choice for sqlite, as sqlite is very different from other databases (https://turriate.com/articles/making-sqlite-faster-in-go, https://crawshaw.io/blog/go-and-sqlite).

 > wrt to go version, why not go straight to 17? the tidy bug seems to be fixed at that version.

@noamnelke was against updating versions if we are not using new features, so i decided not to rush with 1.17.  anyway i think we will want to use 1.18 in near future, as it introduces significant improvements to the language (fuzzing support in stdlib, and generics)