@countvonzero 2022-01-12
> do you remember if database/sql was used in the previous project or a thin wrapper?

yes. we were. we didn't explore other libraries further because core geth devs did not support the change.